script cleanups, added polyid_gals.sh script --Molly
git-svn-id: https://giordano.mit.edu/nvs/mangle/trunk/dev_80bit@125 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ scripts/graphmask.sh
@@ -61,7 +61,9 @@ else
 fi
 if [ -e matlabexit.temp ] || [ ! -e $2 ]; then
     echo <&2 "ERROR: error in matlab plotting script."
-    rm matlabexit.temp
+    if [ -e matlabexit.temp ]; then
+	rm matlabexit.temp
+    fi
     exit 1
 fi
 echo all done!
